Move ConditionalContainerView to it's own file
